LPE-543: Update connectors
